fix(phase-bar): move PhaseProgressBar from Transcript to ContextPanel
- Removed phase computation and PhaseProgressBar from Transcript.tsx
- Added phases prop to ContextPanel component
- PhaseProgressBar now renders in the right sidebar between overview
  and runtime metrics sections
